fix tailwind setup
