+++ src/schemas/movie.ts
@@ -7,10 +7,10 @@ export const MovieSchema = z.object({
   id: z.number(),
   original_language: z.string(),
   original_title: z.string(),
-  overview: z.string().nullable(),
-  popularity: z.number().nullable(),
+  overview: z.string(),
+  popularity: z.number(),
   poster_path: z.string().nullable(),
-  release_date: z.string().nullable(),
+  release_date: z.string(),
   title: z.string(),
   video: z.boolean(),
   vote_average: z.number(),
